refactor(integration): Add general test Cleanup
The test helpers gave two separate cleanup functions. Each package
had to know which subset applies to it, so a new package can miss
one and leak the resource.

Cleanup calls both functions, which are now internal. TestMain of
each package calls only Cleanup, so a later helper needs no change
outside the package.
